Move aggregation tier logic from application_properties to types package
Aggregation tiers, parsing, and resolution are domain logic, not
configuration concerns. Move AggregationTier, DefaultAggregationTiers,
ParseISO8601Duration, ParseAggregationTiers, ResolveAggregationInterval,
and FallbackInterval from application_properties/aggregation_tiers.go
to types/aggregation.go alongside the other aggregation types.

Update all consumers (service, cmd, testharness) to import from types.
Move tests to types/aggregation_tiers_test.go.
Delete application_properties/aggregation_tiers.go.
