+++ sensor_hub/cmd/serve.go
@@ -14,6 +14,7 @@ import (
 	"example/sensorHub/service"
 	"example/sensorHub/smtp"
 	"example/sensorHub/telemetry"
+	"example/sensorHub/types"
 	"example/sensorHub/ws"
 	"fmt"
 	"os"
@@ -124,12 +125,12 @@ func runServe(cmd *cobra.Command, args []string) error {
 		notificationService.CreateNotification(context.Background(), notif, "view_alerts")
 	})
 
-	aggregationTiers, err := appProps.ParseAggregationTiers(appProps.AppConfig.ReadingsAggregationTiers)
+	aggregationTiers, err := types.ParseAggregationTiers(appProps.AppConfig.ReadingsAggregationTiers)
 	if err != nil {
 		return fmt.Errorf("failed to parse aggregation tiers: %w", err)
 	}
 	if aggregationTiers == nil {
-		aggregationTiers = appProps.DefaultAggregationTiers
+		aggregationTiers = types.DefaultAggregationTiers
 	}
 	readingsService := service.NewReadingsService(readingsRepo, mtRepo, aggregationTiers, appProps.AppConfig.ReadingsAggregationEnabled, logger)
 	propertiesService := service.NewPropertiesService(logger)
