+++ sensor_hub/service/readings_service_test.go
@@ -6,7 +6,6 @@ import (
 	"log/slog"
 	"testing"
 
-	appProps "example/sensorHub/application_properties"
 	"example/sensorHub/types"
 
 	"github.com/stretchr/testify/mock"
@@ -21,7 +20,7 @@ import (
 func setupReadingsService() (*ReadingsService, *MockReadingsRepository, *MockMeasurementTypeRepository) {
 	repo := new(MockReadingsRepository)
 	mtRepo := new(MockMeasurementTypeRepository)
-	svc := NewReadingsService(repo, mtRepo, appProps.DefaultAggregationTiers, true, slog.Default())
+	svc := NewReadingsService(repo, mtRepo, types.DefaultAggregationTiers, true, slog.Default())
 	return svc, repo, mtRepo
 }
 
@@ -127,7 +126,7 @@ func TestReadingsService_ServiceGetBetweenDates_UnsupportedFunction(t *testing.T
 func TestReadingsService_ServiceGetBetweenDates_DisabledAggregation(t *testing.T) {
 	repo := new(MockReadingsRepository)
 	mtRepo := new(MockMeasurementTypeRepository)
-	svc := NewReadingsService(repo, mtRepo, appProps.DefaultAggregationTiers, false, slog.Default())
+	svc := NewReadingsService(repo, mtRepo, types.DefaultAggregationTiers, false, slog.Default())
 
 	val := 22.5
 	readings := []types.Reading{
@@ -219,7 +218,7 @@ func TestReadingsService_ServiceGetLatest_Error(t *testing.T) {
 func TestNewReadingsService_ReturnsService(t *testing.T) {
 	repo := new(MockReadingsRepository)
 	mtRepo := new(MockMeasurementTypeRepository)
-	svc := NewReadingsService(repo, mtRepo, appProps.DefaultAggregationTiers, true, slog.Default())
+	svc := NewReadingsService(repo, mtRepo, types.DefaultAggregationTiers, true, slog.Default())
 
 	assert.NotNil(t, svc)
 }
