+++ sensor_hub/testharness/harness.go
@@ -21,6 +21,7 @@ import (
 	"example/sensorHub/notifications"
 	"example/sensorHub/service"
 	"example/sensorHub/smtp"
+	"example/sensorHub/types"
 	"example/sensorHub/ws"
 
 	"github.com/gin-gonic/gin"
@@ -128,7 +129,7 @@ func startServer(sensorURLs []string) (*Env, func(), error) {
 		notificationService.CreateNotification(context.Background(), notif, "view_alerts")
 	})
 
-	tiers := appProps.DefaultAggregationTiers
+	tiers := types.DefaultAggregationTiers
 	readingsService := service.NewReadingsService(readingsRepo, mtRepo, tiers, appProps.AppConfig.ReadingsAggregationEnabled, logger)
 	propertiesService := service.NewPropertiesService(logger)
 	_ = service.NewCleanupService(sensorRepo, readingsRepo, failedRepo, notificationRepo, logger)
