+++ sensor_hub/types/aggregation_tiers_test.go
@@ -1,4 +1,4 @@
-package appProps
+package types
 
 import (
 	"testing"
